move MemberPosition to types
