+++ packages/core/src/utils/interfaces.ts
@@ -412,10 +412,3 @@ export interface InternalSDKLogger extends SDKLogger {
   wsTraffic: (options: WsTrafficOptions) => void
 }
 
-export type MemberPosition =
-  | 'reserved'
-  | `reserved-${number}`
-  | 'standard'
-  | `standard-${number}`
-  | 'off-canvas'
-
